chore: set up git lfs for larger files
I recently realized that even shallow clones of this repo are pretty
large, with a fair amount of that coming from video and image files.

Here I setup the large video in our README to be stored in git lfs
as well as image files for the docs.

This change should be 100% seemless for anyone who has git-lfs installed
and shouldn't affect common workflows.

Also, to reap the benefits in CI, I'm setting up the checkout actions
to more systematically not bring those files as they aren't required
for most CI workloads, with the exception of building docs.

Given the load on CI and how much we checkout the repo there, this
should have a significant impact on speed and bandwidth of CI. I
estimated about 1/2 the size from this handful of files

Note that this change only affects the repo moving forward, meaning
only shallow clones will see a benefit. There's a possibility of doing
this retroactively, but it requires modifying git history and I would
not recommend that.