feat: disable chromium downloads by default in docker-compose
When firing the `docker-compose up` command, which is pretty much the first step
of most developer workflows, we kick up up 2 chromium downloads: one
for PLAYWRIGHT and one for PUPPETEER by default. Those are used for optional features
like alerts/reports/thumbnails/ and for CI.

For convenience, I'm also introducing support for defining
environment variables overrides for docker-compose in `docker/.env-local` while
adding this to `.gitignore`, which developers can use without the
fear/confusion around committing local envrionment specific settings to
the repo.

-----------------

About the core feature in this PR here,
this saves minutes upon firing `docker-compose up`, and reduce confusion
around "what the heck is it doing!?", in a phase where we should just be
pulling and starting docker images.

Most developer workflows don't require either of those, and if it
were the case, we should bake this into the docker image as opposed to
installing during the bootstrap phase. In any case, devs can switch those on
easily by tweaking the env vars specified in `docker/.env`

About PUPPETEER: in https://github.com/apache/superset/pull/22623, I
found that it is used by Applitool in two GHAs:
-
https://github.com/apache/superset/blob/master/.github/workflows/superset-applitool-cypress.yml
-
https://github.com/apache/superset/blob/master/.github/workflows/superset-applitools-storybook.yml

About PLAYWRIGHT, this is what powers alerts and reports as well
thumbnail-generation. This is more common, and can be flipped on
by setting `ENABLE_PLAYWRIGHT=true` in `docker/.env`